+++ src/styled.ts
@@ -2,6 +2,7 @@ import styled, { CreateStyled } from '@emotion/styled';
 
 export interface Theme {
     dir: DocDir;
+    hairlines?: boolean;
 }
 
 const styledWithTheme = (styled as CreateStyled<Theme>);